refactor(bridge): rename class move to moving
