+++ src/test/java/bridge/domain/MapTest.java
@@ -2,7 +2,7 @@ package bridge.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import bridge.util.Move;
+import bridge.util.Moving;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,16 +12,16 @@ public class MapTest {
     @DisplayName("성공 지도 생성")
     @Test
     void testMapSuccess() {
-        map.addMove(Move.UP,false);
+        map.addMove(Moving.UP,false);
         assertThat(map.toString()).isEqualTo("[ O ]\n[   ]");
-        map.addMove(Move.DOWN,false);
+        map.addMove(Moving.DOWN,false);
         assertThat(map.toString()).isEqualTo("[ O |   ]\n[   | O ]");
     }
 
     @DisplayName("실패 지도 생성")
     @Test
     void testMapFail() {
-        map.addMove(Move.UP,true);
+        map.addMove(Moving.UP,true);
         assertThat(map.toString()).isEqualTo("[ X ]\n[   ]");
     }
 }
